moved fakes2aimpl to new file, added unit tests
